% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_norway.R
\docType{data}
\name{border_norway}
\alias{border_norway}
\title{Norwegian border}
\format{A rda file}
\source{
to be added
}
\description{
Norwegian border
}
\references{
to be added
}
\keyword{datasets}

